+++ CHANGELOG.rst
@@ -6,6 +6,21 @@ TODOs
 * pending major release: remove separate prepare step?! initialise in one step during initialisation
 * Numba JIT compilation of utils. line speed profiling for highest impact of refactoring
 * allow input of complex geometries: input coords, and edges separately (polygons are special case)
+* drop simplify parameter
+
+2.4.1 (2022-08-22)
+-------------------
+
+* bugfix: catch the case where no path is possible in the graph in the ``networkx`` A* implementation
+* added speed benchmarks and performance section in the documentation with benchmark results
+
+internal:
+
+* optimisation: checking edges with the biggest angle range first
+* optimisation: skipping visibility checks for the last extremity
+* using optimised point in polygon check algorithm
+* using undirected Graph: The precomputed graph usually makes up the majority of the visibility graph (in comparison to the temporarily added edges for query start and goal nodes) and this precomputed part has to be undirected. Use undirected graph everywhere.
+* added test cases
 
 
 2.4.0 (2022-08-18)